further refine dataviewer
