+++ client/src/components/LibraryNavigator/LibraryModel.ts
@@ -55,7 +55,7 @@ class LibraryModel {
     return new PaginatedResultSet<TableData>(
       async (start: number, end: number) => {
         await this.setup();
-        const limit = end - start;
+        const limit = end - start + 1;
         return await this.retryOnFail(
           async () =>
             await this.dataAccessApi.getRows(
